# Delay retention workload info display after FSRS optimization alerts