--- a/ts/routes/deck-options/FsrsOptions.svelte
+++ b/ts/routes/deck-options/FsrsOptions.svelte
@@ -407,8 +407,7 @@ License: GNU AGPL, version 3 or later; http://www.gnu.org/licenses/agpl.html
 
     :global(.two-line) {
         white-space: pre-wrap;
-        height: calc(2 * 1.5em + 10px);
-        line-height: 1.5em;
+        min-height: calc(2ch + 30px);
         box-sizing: content-box;
         display: flex;
         align-content: center;