When clicking the "Optimize current preset" button in FSRS settings, "The higher this value, the more frequently cards will be shown to you."(using `.two-line`) would experience layout shifts during async operations. The element would initially render with single-line height before the completion alert, then expand to proper two-line height after user dismisses the alert dialog.
![图片](https://github.com/user-attachments/assets/43eff691-3599-4d06-82ae-b24ced9ade3c)
![图片](https://github.com/user-attachments/assets/7aa294c2-5997-4ee0-a32d-3bbc59698e72)
Guess the reason is:
1. High computational issue: Currently using `min-height: calc(2ch + 30px)`, where the `ch` unit is based on the width of the character "0". Layout needs to be recalculated after text updates
2. Rendering Timing Issue: The alert pop-up blocks the main thread, leading to: After state update (`optimized=true`), it cannot immediately re-render, and the browser can only execute the rendering queue after the user clicks "OK".
I Modified CSS styles for `.two-line` class in FSRS settings component.
I haven't tested it, so I don't know if my changes are effective.