@dae 2025-06-08
Unfortunately, this doesn't seem to fix the problem. Changing the setTimeout from 200 to 2000 does, but delay-based "fixes" are bad, as there will always be slower devices that still trigger the problem. I tried inserting an await tick(), but it didn't seem to help. @Luc-Mcgrady in case you're interested