@Luc-Mcgrady 2025-06-08
It's likely caused by the sliding fade in effect of the warning box taking longer than 200ms. A quick google tells me the default is 400ms which would explain it half showing.