@Sunrongguo2008 2025-06-08
It seems that's indeed the case. From the video below, the warning takes about 10 frames to appear. At 48fps, that's roughly 200ms.

https://github.com/user-attachments/assets/3cc28111-4e34-48b3-9ac0-44948a04ea76