@Sunrongguo2008 2025-06-08
I think we can change WORKLOAD_UPDATE_DELAY_MS from 100 to 250 to ensure the warning box appears after alert popups. This resolves timing conflicts where:

    Before:
    100ms → Warning box starts rendering
    200ms → Alert popup appears (interrupting the warning box render)

    After:
    200ms → Alert popup appears first
    250ms → Warning box renders completely