@Luc-Mcgrady 2025-06-08
It might be better to add a 201ms the delay to `optimized = true` (It's not used for anything besides the info box and changing `WORKLOAD_UPDATE_DELAY_MS` would as you could probably tell slow down the rate at which the values update when being changed).

https://github.com/ankitects/anki/blob/c1fc45928dfa3fd12050c2bf254c2e3f33463830/ts/routes/deck-options/FsrsOptions.svelte#L210